rename: 연구실 페이지 상세 컴포넌트 위치 이동
